feat: define habit domain models
