+++ lib/features/habits/domain/domain.dart
@@ -1,0 +1,3 @@
+export 'entities/habit.dart';
+export 'entities/habit_entry.dart';
+export 'value_objects/habit_weekday.dart';
