+++ lib/features/settings/application/providers/notification_settings_provider.dart
@@ -1,0 +1,10 @@
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+
+import '../controllers/notification_settings_controller.dart';
+
+final notificationSettingsProvider =
+    ChangeNotifierProvider<NotificationSettingsController>((ref) {
+      throw UnimplementedError(
+        'Override notificationSettingsProvider before reading it.',
+      );
+    });
