+++ lib/features/settings/application/providers/theme_provider.dart
@@ -1,0 +1,9 @@
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+
+import '../controllers/theme_controller.dart';
+
+final themeControllerProvider = ChangeNotifierProvider<ThemeController>((ref) {
+  throw UnimplementedError(
+    'Override themeControllerProvider before reading it.',
+  );
+});
