--- lib/state/notification_settings_provider.dart
@@ -1,10 +1,0 @@
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-
-import 'notification_settings_controller.dart';
-
-final notificationSettingsProvider =
-    ChangeNotifierProvider<NotificationSettingsController>((ref) {
-  throw UnimplementedError(
-    'Override notificationSettingsProvider before reading it.',
-  );
-});
