--- lib/state/theme_provider.dart
@@ -1,9 +1,0 @@
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-
-import 'theme_controller.dart';
-
-final themeControllerProvider = ChangeNotifierProvider<ThemeController>((ref) {
-  throw UnimplementedError(
-    'Override themeControllerProvider before reading it.',
-  );
-});
